Make it possible to read empty arrays to inspect the contained types.
In the reference implementation this is called type-only mode, although
that one seems to work on any kind of input. I don't see the need for
such generic functionality because you can always throw away the actual
data. That's not expensive compared to parsing overhead.
I might add a possibility to cut short reading of arrays, for just
looking at the types, which needs only one iteration, not all. I don't
currently see a use case for that.

I still want to improve some names (possibly reusing "type-only mode"),
and there are a few TODOs left.