Change cursor API that deals with empty arrays.
It does not only make more sense from outside, it also makes the
implementation look slightly nicer.

This also adds an unfinished test for ReadCursor and WriteCursor to
see how their API would work, and to test the implementation.
For WriteCursor, method stubs are added to make the test compile.