Start implementing some of the things necessary to connect to the bus.
There are some I/O classes, some logic for finding the bus address,
and the beginnings of a d-bus monitor utility to test it all.
The current state is that a local socket connection to the X11 session
bus is established and that's it.
Many extension points are missing - I prefer to add those when needed.