Handle dbus alignment weirdness.
This has already been fixed in serialization.
The problem is that the variable header data does not start on an
8-byte boundary, but ArgumentList expects its first byte to be
perfectly aligned.
So we fudge with data size and starting position in order to fix it up.
When serializing, we postprocess the serialized data, which is possible
while copying it anyway.
When deserializing, it's easiest to prepend a fake int argument. There
are no invalid ints, so there's not much to go wrong. The extra int
just needs to be read and ignored.