Add some alignment-related tests, and make them pass.
The problem was no clear policy regarding alignment of data in the
intermediate data buffer. Alignment in the final buffer is prescribed
by the protocol, but for the intermediate buffer we just use natural
alignment (alignment = data size).
Previously, final alignment was inconsistently applied to the
intermediate buffer, which caused data corruption when reading from it.