Align dict entries after the first.
I discovered this bug by monitoring real data using dselpunk.
It was easy to get this wrong with my API choice to present dicts as
lists of key-value pairs instead of lists (arrays) of dict entries like
the original implementation and the protocol spec.

Unfortunately this bug is hard to test for without sticking serialized
data into the tests, which I don't want to do due to the ugly manual
work involved with that, and the inflexibility of it should any further
problems arise. At least the tests fail when only applying the reader
or only the writer part of this change.