Get out of the connection's client list when we're done with I/O.
Call connection()->removeClient(this).
Found a connection with 17k clients during a cursory look at
performance...

A first version of this patch crashed because
connection()->removeClient(this) was called after
notifyCompletionClient(). In that one, the client can reasonably delete
the completed message. Transceiver does just that when the message is
done writing itself to the connection. With the current object deleted,
connection() returned garbage data so the program crashed.
The solution is the generic one to most callback problems: Call the
callback last and leave the stack frame without touching any further
member variables.
connection()->removeClient(this) mainly changes the return value of
connection(), which is protected, so the completion client shouldn't
see a difference.

Note that the destructor of Message, which includes the one of
IConnectionClient, also removes the Message from the connection's
client list. But we can't *rely* on the Message's client deleting
the Message when it is notified of the Message's completion.

tl;dr: notifyConnectionClient() is free to delete the Message or not,
so we need to handle both cases correctly.