Start doing some introspection data parsing.
AFAICS this is only goin to be used in clients, so there is now a
separate library for clients so that a future server does not need to
link against libtinyxml2.
tinyxml2 is required now because I currently have no plan to make the
client library optional.

There are some changes to support introspection parsing.

Introspection parsing is still very much WIP right now...