d-pointers for ArgumentList and its readers and writers. Also C++11.
They are the core classes, so do them first.
While at it, add copy and move constructors and move assignment
operators as appropriate because they are both easy to implement and
necessary when using d-pointers. Otherwise, deleting the source after
an assignment would delete the shared private object.

The readers and writers are now non-copyable (only movable) which makes
sense semantically. The ArgumentList is only copyable, but the copies
aren't safe yet because there are still pointers that are copied where
the pointed-to data should be copied for real safety.

The ArgumentList tests pass, including valgrind on the tests.