Optimize message header storage a bit.
This might be premature optimization, but it's a pretty decent
ugliness/speed tradeoff. It probably wastes some space if there are
few headers.
The serialization code actually looks a little nicer now IMO, so the
main downside of this is adding more code.

I've also done some thinking about message validation and changed the
message header setters to not return a boolean. It is most convenient
to implement and hopefully to use to just validate the message before
sending. That way we don't need to validate half-built messages.