Rework message ownership when passing from / to Transceiver.
Since Message is noncopyable but movable, there is no reason to use a
unique_ptr - the move semantics are fully built-in with the d-pointer
and all. If we have an indirection, might as well use it.

Also add a cookie parameter to PendingReply, which should often save
its clients a hash / map if they intend to do any routing of their
own. Save a pointer elsewhere in PendingReplyPrivate by exploiting
the fact that transceiver (which supplies a reply) and reply are
never relevant at the same time.