More error propagation for the basic message roundtrip use case.
The idea here is that there is a "workflow" build message, serialize
and validate locally, send, receive reply. For convenience and maybe
safety and performance, it's easiest to check for errors only rarely
and whichever error first occurred is the one reported. It is well
established that the first error is usually the most or only
interesting one.

Different kinds of invalidity of incoming messages are not
distinguished because a sender that cares could probably find out on
its own side what went wrong, and if the sender is malicious there is
really not more to know than that the sender is sending invalid
messages.