Add support for using one DBus connection from multiple threads.
This is of course a relatively large change because some things now
need synchronization.
A maybe uncommon feature is that threads and therefore Transceiver
and EventDispatcher instances that cooperate to make a connection
available in multiple threads may be destroyed in any order without
crashing. Thus destruction of DBus-related objects does not need to
be handled specially in the multi-threaded case. Obviosuly, shutting
down a thread that holds a connection will break connectivity in all
threads that used that one's connection.

The Commutex thing seems like a nice primitive to implement the
necessary linkage between threads, but there is no ready to use
formal or experimental way to check code using it. Ideas for that are
adding appropriate Helgrind / DRD annotations and coming up
with rules that are simple enough to apply during manual code review.

It's an interesting experiment in any case...

This also removes some debug spam that I came across while removing
debug spam added during development of this feature.