Reduce boilerplate for getting arguments in and out of messages.
The nicest part of it is that ArgumentList is now std::move()d out of
ArgumentList::Writer::finish(), so it can be immutable! That is a big
deal, and does away with all the "can't read while writing" logic.
It is simply impossible to write to an existing ArgumentList now.

I really hated that creating a Reader on an ArgumentList was sort of
but not really const because it prevented creation of a Writer on the
same ArgumentList. That is fixed now, hooray!

This change is enabled by move semantics, which I couldn't use when
creating this part of the API because I was only using older standard
C++ at the time.

A sensibly tweaked version of tst_argumentlist:
- repeat main() (excluding "Passed.\n" output) 5000 times
- don't use incomplete input buffers (cf. ArgumentList::NeedMoreData)
- exclude for fairness the readerWriterExclusion() test because in the
  new API we don't need and it and couldn't even implement it.
- move writer.finish() to after the big case statement in
  doRoundtripForReal()
runs in about 3.5 seconds after this change - it took 5 seconds before.
That is clearly an implausible improvement from that kind of change :)
In real use cases it should still bring a small performance gain.