Revert "HACK: Fix a crash. This should not be necessary because...
...the ArgumentList is copied."

The "hack" worked fine, but I'd rather do it the obvious way.

Presumably the problem was fixed in the large Transceiver rework for
multithreading. There is no crash or Valgrind problem in any of the
tests, so it should be fine. It would be nice to find out what the
problem was, but well... there are more useful things to do.

This reverts commit 5c89daff101ae064b38101571e10941cc230bb47.