Allow reading nil arrays with ...::Reader::readPrimitiveArray().
Writing was already possible, that part was rather trival.

There is a notable fix of a latent bug in this, arrayInfo.dataEnd is
now initialized even when reading a nil array. We just never read
the uninitialized value before. It was well possible to ignore the
bad value in readPrimitiveArray() but I'd rather not have a bad
value in the first place.