Memory optimizations in Arguments.
Put a thread-local (no locks) cache for common allocations in front
of malloc and add a few more reserve()s.
For the allocation cache, there is the "problem" of moving an
allocation to another thread. But it's not really a problem because
malloc will handle the eventual free() in the new thread, which it
can already do.

The somewhat large diff is mostly due to moving code around which
was necessary.