Make exported symbols "protected".
That means that they cannot be interposed by another (usuall preloaded)
library, which does not seem to be a big loss. The good thing about
it is that exported functions can now be inlined.

This change is motivated by looking at assembly generated with the last
change and seeing some lack of inlining... you can't inline functions
that could be overridden using dynamic loading tricks.