Arguments::prettyPrint(): add a special case for byte arrays.
As the general purpose raw data format, printing each byte on its own
line uses an extreme amount of screen real estate. Change that.

This is a good use case for readPrimitiveArray(). Since we only want to
handle byte arrays specially, and both readPrimitiveArray() and
"opening" an array the normal way are not undoable, we need a method
to peek into the array to check for the desired type (bytes).
Add the method peekPrimitiveArray() for that.