Arguments::Reader: Implement skipping of all aggregate types.
With extensive tests.
This facilitates converting / consuming aggregate-type arguments
lazily: copy the reader at the beginning of the aggregate, then
skip. Of that scheme, the copy and resume part is currently not
tested.

This also fixes Reader::begin(SkipIfEmpty) with arrays that really
are empty. The signature position was off by one afterwards.
The just added test finally stresses that functionally.