WIP: Arguments::{Reader,Writer}: Remove Next{Array,Dict}Entry states.
They were meant to be self-documentation and a small help in certain
cases, but they are not required for the API to work and they also
make things a few percent slower in the average case.
The real reason for this change is compatibility with QDBus API,
though.

WIP because:
- need to fix pretty-printing where NextDictEntry was previously used
- found some missing input validation that should be added