Arguments::Writer: make it copyable.
Not that useful normally, but QDBusArgument unfortunately must
be copyable in any state.
Only a tiny fraction of copied data needs to be fixed up due to
the change to store variant signatures "inline" in m_data, and
the resulting ability to use positions relative to m_data for
signatures, which works best in that constellation anyway.