Allow to restart an array / dict as empty if not yet written to.
This is needed in client code where the number of elements is not
known at the time beginArray/Dict() is called, but there is
enough information for a late fix-up. As happens to be the case
in QtDBus.
This is just a little hack to solve a specific problem and should
not be the first choice when writing possibly empty arrays.