Rework handling of "state" in Arguments::Writer.
It was very similar to Reader and that didn't make sense. For Writer,
the user-visible state either says "you must write xyz next" or "you
can write whatever you want next" or "something went wrong, this
instance is hosed". It makes no sense to have state equal the last
thing written, which happened most of the time.

This part of the API is almost never needed, but at least it makes
sense now.