Allow patching of serial on an already serialized Message
In some situations, serializing Messages explicitly before sending
can be used to better distribute CPU usage. But the serial is often
(necessarily, because the serial is specific to the I/O channel)
assigned just before sending. This change achieves that the latter
does not destroy the performance advantage of the former.