Use new interface when old monitoring interface is disabled
The old interface depends on configuration, which varies by
distribution. For reasons that aren't quite clear from reading
the config file, the old interface doesn't work anymore on
Kubuntu (18.10) and I've been using a similar patch for some
time - now fix it for everyone.

The shared code is IMO too much for straight duplication but also
does not fit well into one of the libraries, so just make it an
included snippet.