Local and IP (on Unix) sockets: suppress SIGPIPE
SIGPIPEs made the connection/server test flaky.

Sending on a closed socket (usually closed by remote, or we'd
notice before trying to send) provokes a SIGPIPE without these
flags. But signal handlers are kind of broken by design, and in
any case we can simply look at the return value and errno of
send() or equivalent.