Limit scope of StateChanger to not cross callback invocations
Otherwise, the following can happen:
- StateChanger is instantiated
- StateChanger::setNewState() is called (or the corresponding
  convenience constructor is used) - it remembers the current
  state as m_oldState.
- API client controlled callback is invoked, which calls something
  that changes connection state, e.g. Connection::close().
- StateChanger is destroyed, at which point its m_oldState is
  outdated due to the change made by the callback. It could signal
  a bogus state change.

It is *generally* a good idea to instantiate StateChanger early to
make it less likely to forget to signal state changes, and so that
it is destroyed late during function exit, which follows the
general rule to invoke callbacks as late as possible. But the
above mentioned problem limits that.