Fix adding and removing timers while in their callback
Immediately after invoking the callback, the event dispatcher
checks whether to re-add the timer and when it should trigger.
Calling addTimer() or removeTimer() *while in the callback* is not
necessary and doing so actually causes a mess. Just let the checks
after the callback invocation do their work.

This fixes a timer serial overflow warning in tst_timer_slow
due to the same timer getting added *many* times to the event
dispatcher's timer map.