Fix handling of timers inserted from inside a timer callback
Basically, due time = 0 constitutes a staging area in the timer
map for timers inserted during event loop iteration that should
trigger earliest in the next iteration (and would trigger in the
current iteration due to interval = 0). When the next iteration
comes, transfer all timers from the staging area to the active
area before triggering any timers.
This is more robust and doesn't seem to have cases in which it
doesn't work (except monotonic time being exactly zero...).
This (finally!) un-flakifies the testAddInTrigger() test.
Also tighten the comments a little.