Handle all connection setup errors in callbacks
As opposed to code where we "just know" that we are currently
cranking the connection setup state machine, especially
Connection::waitForConnectionEstablished().
This both expands actually doing anything about most such errors
and unifies how to do it.
Also introduce a new error code range for errors that are most
closely associated with the Connection. Some existing ones fit in
much better in that area. For example, there is no such thing as
too many file descriptors except when trying to send a Message over
a Connection that can only transmit a certain number per datagram.