Add support for TCP connections to other hosts, not just loopback
Somewhat limited (no DNS hostname resolution) and untested except
with loopack as before. Apart from the set of possible errors,
connect() et al work the same way no matter on which computer an
IP address is located - so error handling (when a remote connection
is actually used) is where this is most likely to cause trouble.