Fix a crash in (and clean up) SelectEventPoller for Unix-likes
It was iterating over EventDispatcherPrivate::m_ioListeners and
(indirectly) invoking methods on the listeners that, in some cases,
resulted in them getting removed from m_ioListeners. So the
iteration proceeded with an invalidated iterator.
While investigating the situation, I also found unused data members
in SelectEventPoller that I removed.
Fix the problem by iterating over the select result sets in the
most simple and stupid way, which is probably not all that slow
given the cache-friendly memory access pattern and the typically
small value of FD_SETSIZE (1024 on Linux).