Rework read() and write() to have only one implementation each...
- with optional file descriptor support (at least according to the
  signature, it still can't work with TCP/IP) - this is just to
  reduce code duplication
- with scatter-gather ("multi buffer") support in write()

The purpose of the latter is to optimize writing messages with a
large payload. It saves copying together message header and
payload just to have a single buffer to write. IpSocket, which
doesn't have support implemented for now, will need to fall back to
making multiple syscalls when passed multiple buffers.