Use multi-buffer send capability in Message
This saves copying payload data (if any), which speeds up sending
"large, dumb" data (like byte arrays) by a few percent to factor
6 (for a direct connection without DBus daemon) on supported
transports, which is only local sockets currently.
The difference in direct mode is crazy. My best guess is that it
exercises some unknown-to-me behavior in the I/O stack because
there is a huge difference in "system" CPU time as measured by the
time command.
The worse performance in direct mode before this commit is not a
regression from recent changes in I/O.

The fallback support in IpSocket needs to send the data with two
send() calls, which absolutely ruins latency (and thus transfer
rate) of a ping test on Linux - so, in case a transport doesn't
really implement multi-buffer, do it the old way. I have even tried
setting MSG_NODELAY but only on the last send(), which didn't help.
Worth noting that all POSIX systems and even Winsock do support
something like iovec for TCP/IP, it's just not implemented here.
And, of course, it requires leaving the almost-POSIX happy path
in Winsock and using Windows-only API.

The performance differences mentioned in the first and the second
paragraph *seem* to be unrelated. In the first paragraph, the CPU
is busy. In the second paragraph, it looks more like a scheduling
issue because the CPU load is only about 30%.

I have removed serializeAndView() because that one became difficult
to do well (fast and easy to use) and it has no callers.